requirements.txt
anyio==3.6.1
asgiref==3.4.1
attrs==21.4.0
bidict==0.22.0
cachelib==0.7.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==35.0.0
cycler==0.11.0
Deprecated==1.2.13
filelock==3.7.0
Flask==2.1.2
Flask-Login==0.6.1
flask-mongo-sessions==0.2.1
Flask-PyMongo==2.3.0
Flask-Session==0.4.0
Flask-SocketIO==5.2.0
fonttools==4.33.3
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.3
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==3.2.0
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.2
numpy==1.22.4
packaging==21.3
pandas==1.4.2
phonenumbers==8.12.0
Pillow==9.1.1
pycparser==2.21
pycryptodome==3.10.4
PyJWT==2.3.0
pymongo==4.1.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-engineio==4.3.2
python-socketio==5.6.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-file==1.5.1
rfc3986==1.5.0
Shapely==1.8.2
six==1.16.0
sniffio==1.2.0
supertokens-python==0.8.1
tldextract==3.1.0
typing_extensions==4.1.1
ua-parser==0.10.0
urllib3==1.26.9
user-agent-parser==0.1.1
user-agents==2.2.0
Werkzeug==2.0.3
wrapt==1.14.1
